Add calendar view for admins
