+++ app/Http/Controllers/AdminController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Message;
+use App\ScheduledMessage;
 use App\Subscriber;
 use Illuminate\Http\Request;
 
@@ -11,6 +11,7 @@ class AdminController extends Controller
     public function index(Request $request)
     {
         return view('admin.index', [
+            'sendDates' => ScheduledMessage::pluck('send_at'),
             'subscriberCount' => Subscriber::all()->count(),
             'subscribersThisWeek' => Subscriber::newThisWeek()->count(),
             'subscribersEN' => Subscriber::all()->where('locale', 'en')->count(),
