+++ resources/css/app.css
@@ -7,5 +7,6 @@
 @import 'link';
 @import 'btn';
 @import 'bubbles';
+@import 'flatpickr';
 
 @import 'tailwindcss/utilities';
