+++ resources/js/dropdown.js
@@ -1,0 +1,14 @@
+export default () => {
+    return {
+        show: false,
+        open() {
+            this.show = true;
+        },
+        close() {
+            this.show = false;
+        },
+        isOpen() {
+            return this.show === true;
+        }
+    };
+};
