+++ resources/views/components/character-count.blade.php
@@ -20,56 +20,3 @@
 @error($name)
     <p class="mt-2 text-sm text-red-600" id="{{ $id }}-error">{{ $message }}</p>
 @enderror
-
-<script>
-    function messageBox () {
-        return {
-            message: '',
-            stats: {
-                messages: 0,
-                remaining: 160
-            },
-            fetchMessageCount: debounce(function () {
-                if (! this.message.length) {
-                    this.resetStats()
-                    return
-                }
-
-                axios.get('/api/message-count', {
-                    params: {
-                        message: this.message
-                    }
-                })
-                .then(response => {
-                    this.stats = response.data
-                })
-                .catch(error => {
-                    console.error(error)
-                })
-            }, 1000),
-
-            get characterCount() {
-                return this.message.length + ' ' + ' character' + (this.message.length !== 1 ? 's' : '')
-            },
-
-            get messageCount() {
-                return this.stats.messages + ' messages'
-            },
-
-            get messageCharactersRemaining() {
-                return this.stats.remaining + ' characters remaining'
-            },
-
-            get messageStats() {
-                return this.characterCount + ' | ' + this.messageCount + ' | ' + this.messageCharactersRemaining
-            },
-
-            resetStats() {
-                this.stats = {
-                    messages: 0,
-                    remaining: 160
-                }
-            }
-        }
-    }
-</script>
